+++ Classes/Configuration/ListViewConfigurationCheck.php
@@ -143,7 +143,7 @@ class ListViewConfigurationCheck extends AbstractFrontEndConfigurationCheck
     }
 
     /**
-     * @deprecated #1788 will be removed in seminars 5.0
+     * @deprecated #1788 will be removed in seminars 6.0
      */
     private function checkOmitDateIfSameAsPrevious(): void
     {