+++ CHANGELOG.md
@@ -8,7 +8,6 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 
 ### Changed
-- Improve some type annotations (#2495, #2497)
 
 ### Deprecated
 - Update some `@deprecated` annotations (#2498)
@@ -16,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Improve some type annotations (#2495, #2497)
 - Drop the legacy registration form parameters from `DefaultController::getLoginLink` (#2500)
 - Use `Connection` instead of `PDO` for the named param types (#2496)