+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Convert most classes that get injected into singletons (#2503)
 - Improve some type annotations (#2495, #2497)
 - Drop the legacy registration form parameters from `DefaultController::getLoginLink` (#2500)
 - Use `Connection` instead of `PDO` for the named param types (#2496)